refactor(tests): move conftest
